[TST] Consolidate shared package version helpers
