+++ chromadb/test/property/test_cross_version_persist.py
@@ -1,10 +1,7 @@
 from multiprocessing.connection import Connection
-import sys
 import os
 import shutil
-import subprocess
 import tempfile
-from types import ModuleType
 from typing import Generator, List, Tuple, Dict, Any, Callable, Type
 from hypothesis import given, settings
 import hypothesis.strategies as st
@@ -28,6 +25,11 @@ import re
 import multiprocessing
 from chromadb.config import Settings
 from chromadb.api.client import Client as ClientCreator
+from chromadb.test.utils.cross_version import (
+    switch_to_version,
+    install_version,
+    get_path_to_version_install,
+)
 
 # Minimum persisted version we support, and other substantial change versions
 # 0.4.1 is the first version with persistence
@@ -157,75 +159,6 @@ def version_settings(request) -> Generator[Tuple[str, Settings], None, None]:
         shutil.rmtree(data_path, ignore_errors=True)
 
 
-def get_path_to_version_install(version: str) -> str:
-    return base_install_dir + "/" + version
-
-
-def get_path_to_version_library(version: str) -> str:
-    return get_path_to_version_install(version) + "/chromadb/__init__.py"
-
-
-def install_version(version: str) -> None:
-    # Check if already installed
-    version_library = get_path_to_version_library(version)
-    if os.path.exists(version_library):
-        return
-    path = get_path_to_version_install(version)
-    install(f"chromadb=={version}", path)
-
-
-def install(pkg: str, path: str) -> int:
-    # -q -q to suppress pip output to ERROR level
-    # https://pip.pypa.io/en/stable/cli/pip/#quiet
-    print("Purging pip cache")
-    subprocess.check_call(
-        [
-            sys.executable,
-            "-m",
-            "pip",
-            "cache",
-            "purge",
-        ]
-    )
-    print(f"Installing chromadb version {pkg} to {path}")
-    return subprocess.check_call(
-        [
-            sys.executable,
-            "-m",
-            "pip",
-            "-q",
-            "-q",
-            "install",
-            pkg,
-            "--no-binary=chroma-hnswlib",
-            "--target={}".format(path),
-        ]
-    )
-
-
-def switch_to_version(version: str) -> ModuleType:
-    module_name = "chromadb"
-    # Remove old version from sys.modules, except test modules
-    old_modules = {
-        n: m
-        for n, m in sys.modules.items()
-        if n == module_name
-        or (n.startswith(module_name + "."))
-        or n in VERSIONED_MODULES
-        or (any(n.startswith(m + ".") for m in VERSIONED_MODULES))
-    }
-    for n in old_modules:
-        del sys.modules[n]
-
-    # Load the target version and override the path to the installed version
-    # https://docs.python.org/3/library/importlib.html#importing-a-source-file-directly
-    sys.path.insert(0, get_path_to_version_install(version))
-    import chromadb
-
-    assert chromadb.__version__ == version
-    return chromadb
-
-
 class not_implemented_ef(EmbeddingFunction[Documents]):
     def __call__(self, input: Documents) -> Embeddings:
         assert False, "Embedding function should not be called"
@@ -239,7 +172,7 @@ def persist_generated_data_with_old_version(
     conn: Connection,
 ) -> None:
     try:
-        old_module = switch_to_version(version)
+        old_module = switch_to_version(version, VERSIONED_MODULES)
         system = old_module.config.System(settings)
         api = system.instance(api_import_for_version(old_module, version))
         system.start()
